move get_data to functions #64
